Fix bug in jenkins script

+++ scripts/create_virtualenv.sh
@@ -27,6 +27,12 @@ fi
 
 VENV="${PWD}/venv_tmlibrary"
 
+# Some packages require the latest pip version
+if [[ ! $(pip -V) =~ "7.1.2" ]]; then
+    echo "requires pip version 7.1.2"
+    exit 101
+fi
+
 # Remove previous environment
 if [ -d $VENV ]; then
   rm -r VENV
@@ -44,7 +50,7 @@ pip install -r requirements-2.txt
 
 # NOTE: dependencies in requirements-Darwin-3.txt should be installed globally
 # via apt-get
-ln -s /usr/local/lib/python2.7/site-packages/gi $VENV/lib/python2.7/site-packages/gi
+ln -s /usr/lib/python2.7/dist-packages/gi $VENV/lib/python2.7/site-packages/gi
 ln -s /usr/lib/python2.7/dist-packages/cv2.so $VENV/lib/python2.7/site-packages/cv2.so
 ln -s /usr/lib/python2.7/dist-packages/lxml $VENV/lib/python2.7/site-packages/lxml
 ln -s /usr/lib/python2.7/dist-packages/rpy2 $VENV/lib/python2.7/site-packages/rpy2
